Introduction of the sparseData module and SparseCollection
This also removes the `sortAsync` and 'fetchJumbo' modules.
It also cleans up a lot of our worker management functionality.
It also provides a fancy example that runs separately from the docs
As a side effect it fixes Issue #19.

+++ src/Boss.js
@@ -6,7 +6,15 @@ var when = require('when');
 /**
  * This object provides an interface to a worker that communicates via promises.
  * Still conflicted about whether this should be an external module or not
- * @param options
+ * @param options which includes:
+ *   o fileLocation (required):  The location of the Worker JS file to load
+ *   o Worker (required): The Worker constructor to use.  Typically will be window.Worker
+ *     unless writing tests
+ *   o autoTerminate (optional):  If boolean false, the worker will never be terminated.  If boolean true,
+ *     the worker will be terminated immediately.  If a number, the worker will be terminated after that many
+ *     milliseconds.  Note that the worker will always be recreated when necessary (i.e. when calling
+ *     <code>boss.makePromise(...)</code>.  This defaults to 1000, meaning a worker will be terminated if it is not
+ *     used for one second.
  * @constructor
  */
 function Boss(options) {
@@ -34,12 +42,12 @@ Boss.prototype = {
      * Get a promise that will be resolved when the worker finishes
      * @param details Details for the method call:
      *   o method (required) The name of the method to call
-     *   o data (optional) Any data that should be passed to the worker method
-     *     you are calling
+     *   o argument (optional) The single argument that will be passed to the
+     *     worker method you are calling
      * @return A Promise that will be resolved or rejected based on calling
      *   the method you are calling.
      */
-    promise: function(details) {
+    makePromise: function(details) {
         if (!_.isString(details.method)) {
             throw new Error("Must provide 'method'");
         }
@@ -74,6 +82,8 @@ Boss.prototype = {
                 reject(err);
             };
 
+            // TODO:  if details.argument is an easily measurable payload (i.e. a long string),
+            // use an ArrayBuffer to speed the transfer.
             worker.postMessage(details);
         }).finally(function() {
             // Set a timeout to terminate the worker if it is not used quickly enough
@@ -87,7 +97,9 @@ Boss.prototype = {
      */
     terminate: function() {
         if(this.worker) {
-            this.worker.terminate();
+            if (_.isFunction(this.worker.terminate)) {
+                this.worker.terminate();
+            }
             this.worker = null;
         }
     },
